Fixed coarse increment decimals causing exceptions
git-svn-id: http://svn.3splooges.com/romraider-arch/trunk@52 d2e2e1cd-ba16-0410-be16-b7c4453c7c2d

+++ src/enginuity/maps/Table.java
@@ -535,7 +535,7 @@ public abstract class Table extends JPanel implements Serializable {
         return frame;
     }
     
-    public void increment(int increment) {
+    public void increment(double increment) {
         if (!isStatic) {
             for (int i = 0; i < data.length; i++) {
                 if (data[i].isSelected()) data[i].increment(increment);
